+++ src/CheckBox/storybook/CheckBox.stories.tsx
@@ -19,11 +19,36 @@ WithError.args = {
   children: 'I disagree',
   error: true,
 }
+WithError.parameters = {
+  a11y: {
+    config: {
+      rules: [
+        {
+          //disabling these until we sync with design on this
+          id: 'color-contrast',
+          enabled: false,
+        },
+      ],
+    },
+  },
+}
 
 export const WithErrorLabel = Template.bind({})
-
 WithErrorLabel.args = {
   children: 'I disagree',
   error: true,
   errorMsg: 'Something has gone wrong',
+}
+WithErrorLabel.parameters = {
+  a11y: {
+    config: {
+      rules: [
+        {
+          //disabling these until we sync with design on this
+          id: 'color-contrast',
+          enabled: false,
+        },
+      ],
+    },
+  },
 }