+++ src/Dropdown/storybook/Dropdown.stories.tsx
@@ -46,6 +46,7 @@ export const Default = Template.bind({})
 Default.args = {
   id: 'days',
   list: days,
+  label: 'Days dropdown',
   placeholder: 'Placeholder',
 }
 
@@ -54,8 +55,22 @@ export const DefaultFallback = Template.bind({})
 DefaultFallback.args = {
   id: 'days',
   list: days,
+  label: 'Days dropdown',
   fallbackStyle: true,
   placeholder: 'Placeholder',
 }
 
 export const Collection = CollectionPage.bind({})
+Collection.parameters = {
+  a11y: {
+    config: {
+      rules: [
+        {
+          //disabling these until we sync with design on this
+          id: 'color-contrast',
+          enabled: false,
+        },
+      ],
+    },
+  },
+}