+++ src/NumberInput/storybook/NumberInput.stories.tsx
@@ -35,6 +35,19 @@ DefaultFallback.args = {
 }
 
 export const Collection = CollectionPage.bind({})
+Collection.parameters = {
+  a11y: {
+    config: {
+      rules: [
+        {
+          //disabling these until we sync with design on this
+          id: 'color-contrast',
+          enabled: false,
+        },
+      ],
+    },
+  },
+}
 
 const WorkingExampleTemplate = () => <Container />