+++ src/SearchInput/components/EmptyResults.tsx
@@ -1,0 +1,23 @@
+import React, { FC, ReactNode, useEffect } from 'react'
+
+interface Props {
+  onNotFound?: (searchTerm: string) => void
+  searchTerm: string
+  notFoundComponent?: ReactNode
+}
+
+export const EmptyResults: FC<Props> = ({
+  onNotFound,
+  searchTerm,
+  notFoundComponent,
+}) => {
+  useEffect(() => {
+    onNotFound?.(searchTerm)
+  }, [])
+
+  if (notFoundComponent) {
+    return <li>{notFoundComponent}</li>
+  }
+
+  return <li>No results</li>
+}